requirements.txt
streamlit
pandas
python-dotenv
google-genai
requests
numpy
scikit-learn
xgboost
joblib
plotly
pydantic
